+++ src/components/Link/Link.js
@@ -6,16 +6,8 @@ import { Link as GatsbyLink } from "gatsby"
 import colors from "../../theme/colors"
 import { showCustomCssDeprecationMessage } from "../../utils/maintenance/deprecationMessages"
 
-const Link = ({
-  children,
-  href,
-  target,
-  to,
-  customCss,
-  variant = `DEFAULT`,
-  ...rest
-}) => {
-  showCustomCssDeprecationMessage(customCss)
+const Link = ({ children, href, target, to, variant = `DEFAULT`, ...rest }) => {
+  showCustomCssDeprecationMessage(rest.customCss)
 
   const baseStyles = {
     alignItems: `center`,
@@ -46,7 +38,7 @@ const Link = ({
         ...baseStyles,
         ...styles[variant],
       },
-      customCss,
+      rest.customCss,
     ],
     ...rest,
   }
