+++ src/components/Toast/constants.ts
@@ -1,1 +1,4 @@
 export type ToastTones = "SUCCESS" | "DANGER"
+
+export const DEFAULT_TIMEOUT = 5000
+export const DEFAULT_TONE: ToastTones = `SUCCESS`
