--- src/components/Toast/index.js
@@ -1,2 +1,0 @@
-export { ToastProvider, ToastConsumer } from "./ToastContext"
-export * from "./hooks"
